Replace regexp match with stdlib path.Split() parsing.
This is simpler and more robust. The regexp solution, for instance,
served up 'index.html' if you ask for 'index.htmlnotreally'.

The early 404 if the regexp doesn't match is not necessary - the
swaggerFiles handler will return 404 if the filename doesn't match any
swaggerUI content.

Setting the prefix on the handler is also not necessary, as we are
already passing in the filename, so the prefix doesn't need to be
stripped.