Fix link to Swagger spec in Swagger UI by reworking URL config.
This makes it more clear that this config option should only be set if
the Swagger spec document is not expected to be hosted by this
middleware but instead by an external website or service. It also fixes
the link to the JSON document under the base path on the Swagger UI.

Also set the swaggerFiles handler prefix again, as this is required.